Remove various build systems in favour of CMake
This removes support for GNU autotools, as well as a whole bunch of
IDE project files.
The plan is to consolidate effort behind a single build system rather
that trying to maintain a bunch of them. And CMake seems to be the
best supported across the various target platforms.

The old 'projects' directory contained a bunch of language bindings.
Those have been moved into a new directory, 'bindings'.