il_devil.cpp: Rename mips to mipmaps
Something in my MIPS toolchain defines a constant `mips`, resulting in
compilation error here. My toolchain comes from buildroot so it's likely
that other users may experience the same issue.

I'm hoping to fix the root cause later on my side, but for now the
easiest seems to be to rename the variable in DevIL.